modularity added to the code
